pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "Resident infrastructure agent — observes k8s clusters, remembers, and reports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastharness[deepagents]>=1.1.0",
    "mem0ai>=0.1.0",
    "kubernetes>=31.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pulse"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
